requirements.txt
absl-py==2.1.0
tensorboardX==2.6.2.2
tqdm==4.66.1
imageio==2.34.0
scipy==1.12.0
scikit-learn==1.4.1.post1
matplotlib==3.8.0
ray==2.9.1
pytorch-fid==0.3.0
torch-pruning